requirements.txt
django
gunicorn
django-heroku
requests
djangorestframework
django_oauth2_provider
